+++ Views/AddAddonDialog.Designer.cs
@@ -19,31 +19,32 @@ namespace Teron_Addon_Manager
 
         private void InitializeComponent()
         {
+            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(AddAddonDialog));
             promptLabel = new Label();
             urlTextBox = new TextBox();
             okButton = new Button();
             cancelButton = new Button();
             SuspendLayout();
-            //
+            // 
             // promptLabel
-            //
+            // 
             promptLabel.AutoSize = true;
             promptLabel.Location = new Point(12, 15);
             promptLabel.Name = "promptLabel";
-            promptLabel.Size = new Size(220, 15);
+            promptLabel.Size = new Size(169, 15);
             promptLabel.TabIndex = 0;
             promptLabel.Text = "Addon page or download URL:";
-            //
+            // 
             // urlTextBox
-            //
+            // 
             urlTextBox.Anchor = AnchorStyles.Top | AnchorStyles.Left | AnchorStyles.Right;
             urlTextBox.Location = new Point(12, 35);
             urlTextBox.Name = "urlTextBox";
             urlTextBox.Size = new Size(396, 23);
             urlTextBox.TabIndex = 1;
-            //
+            // 
             // okButton
-            //
+            // 
             okButton.Anchor = AnchorStyles.Bottom | AnchorStyles.Right;
             okButton.Location = new Point(252, 75);
             okButton.Name = "okButton";
@@ -51,9 +52,9 @@ namespace Teron_Addon_Manager
             okButton.TabIndex = 2;
             okButton.Text = "OK";
             okButton.UseVisualStyleBackColor = true;
-            //
+            // 
             // cancelButton
-            //
+            // 
             cancelButton.Anchor = AnchorStyles.Bottom | AnchorStyles.Right;
             cancelButton.DialogResult = DialogResult.Cancel;
             cancelButton.Location = new Point(333, 75);
@@ -62,25 +63,27 @@ namespace Teron_Addon_Manager
             cancelButton.TabIndex = 3;
             cancelButton.Text = "Cancel";
             cancelButton.UseVisualStyleBackColor = true;
-            //
+            // 
             // AddAddonDialog
-            //
+            // 
             AcceptButton = okButton;
-            CancelButton = cancelButton;
             AutoScaleDimensions = new SizeF(7F, 15F);
             AutoScaleMode = AutoScaleMode.Font;
+            CancelButton = cancelButton;
             ClientSize = new Size(420, 114);
             Controls.Add(promptLabel);
             Controls.Add(urlTextBox);
             Controls.Add(okButton);
             Controls.Add(cancelButton);
             FormBorderStyle = FormBorderStyle.FixedDialog;
+            Icon = (Icon)resources.GetObject("$this.Icon");
             MaximizeBox = false;
             MinimizeBox = false;
             Name = "AddAddonDialog";
             StartPosition = FormStartPosition.CenterParent;
             Text = "Add Addon";
             ResumeLayout(false);
+            PerformLayout();
         }
     }
 }